moved id() into utils file
