+++ src/util/utils.js
@@ -1,0 +1,20 @@
+// utils..
+
+
+function makeId() {
+    function s4() {
+        return Math.floor((1 + Math.random()) * 0x10000)
+            .toString(16)
+            .substring(1)
+    }
+
+    return s4() + s4() + '-' + s4() + '-' + s4() + '-' +
+        s4() + '-' + s4() + s4() + s4()
+}
+
+
+
+
+export {
+    makeId,
+}
